+++ app/models/invitation.rb
@@ -8,7 +8,7 @@ class Invitation < ActiveRecord::Base
   belongs_to :invitee, class_name: 'User', inverse_of: :invitations
   belongs_to :inviter, class_name: 'User', inverse_of: :invitations_from_me
   belongs_to :actor, class_name: 'User'
-  before_create :assign_to_latest_decision
+  before_create :assign_to_draft_decision
 
   scope :where_email_matches,
         ->(email) { where('lower(email) = lower(?) OR lower(email) like lower(?)', email, "%<#{email}>") }
@@ -46,7 +46,7 @@ class Invitation < ActiveRecord::Base
 
   def self.find_uninvited_users_for_paper(possible_users, paper)
     invited_users = where(
-      decision_id: paper.decisions.latest.id,
+      decision_id: paper.draft_decision.id,
       state: ["invited", "accepted", "declined"]
     ).includes(:invitee).map(&:invitee)
     possible_users - invited_users
@@ -93,8 +93,8 @@ class Invitation < ActiveRecord::Base
 
   private
 
-  def assign_to_latest_decision
-    self.decision = paper.decisions.latest
+  def assign_to_draft_decision
+    self.decision = paper.draft_decision
   end
 
   def add_authors_to_information(invitation)